src/distance_selector/dists/src/.history/DISTS_pytorch/video_requirements_20250921152105.txt
# Requirements for video_dists_analyzer.py
opencv-python>=4.5.0
torch>=1.0.0
torchvision>=0.2.0
matplotlib>=3.3.0
numpy>=1.19.0
Pillow>=8.0.0